Moved Split to upper level.
