+++ src/main/java/org/ltr4l/trainers/LambdaMartTrainer.java
@@ -16,7 +16,7 @@
 package org.ltr4l.trainers;
 
 import org.ltr4l.boosting.*;
-import org.ltr4l.boosting.RegressionTree.Split;
+import org.ltr4l.boosting.Split;
 import org.ltr4l.nn.Activation;
 import org.ltr4l.query.Document;
 import org.ltr4l.query.Query;
